separate DigiAssets classes created
