+++ BRAssetData.c
@@ -1,0 +1,9 @@
+//
+//  BRAssetData.c
+//  DigiByte
+//
+//  Created by Julian Jäger on 05.10.19.
+//  Copyright © 2019 DigiByte Foundation NZ Limited. All rights reserved.
+//
+
+#include "BRAssetData.h"
